// modules/xgui/src/controls/CheckBox.d.ts
/// <reference path="../../typings/angular2/angular2.d.ts" />
/**
 * Created by Nidin Vinayakan on 02-10-2015.
 */
import { EventEmitter } from 'angular2/angular2';
export declare class CheckBox {
    click: EventEmitter;
    elements: any;
    private _label;
    constructor();
    label: string;
}